pyproject.toml
[tool.poetry]
name = "python-web"
version = "0.1.0"
description = ""
authors = ["Sudarenko Sergey <Vipgaz2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
django-environ = "^0.11.2"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
